Modularize node web APIs used for mocking
